Struct pikadick_system_info::CacheContext

pub struct CacheContext { /* private fields */ }

A context for caching data related to information queries.

A system may occasionally give more data than requested when asked. Using this allows the use of that data without fetching it again. A new context should be created if new data is required, as it will happily reuse old data if it has it.

Implementations

impl CacheContext

pub fn new() -> Self

Make a new CacheContext.

pub fn get_boot_time(&self) -> Result<SystemTime, Error>

Get the boot time.

Blocking

This is NOT blocking.

pub fn get_uptime(&self) -> Result<Duration, Error>

Get the computer uptime.

Blocking

This is NOT blocking.

pub fn get_hostname(&self) -> Result<String, Error>

Get the hostname.

Blocking

This is NOT blocking.

pub fn get_architecture(&self) -> Result<Option<Arch>, Error>

Get the arch.

Returns

This returns None if the arch is unknown.

Blocking

This is NOT blocking.

pub fn get_system_name(&self) -> Result<Option<String>, Error>

Get the system name, or the name of the operating system.

Blocking

This is NOT blocking.

pub fn get_system_version(&self) -> Result<String, Error>

Get the operating system version.

Blocking

This is NOT blocking.

pub fn get_total_memory(&self) -> Result<u64, Error>

Get the total amount of memory in the computer, in bytes.

Blocking

This is NOT blocking.

pub fn get_available_memory(&self) -> Result<u64, Error>

Get the available amount of memory in the computer, in bytes.

Blocking

This is NOT blocking.

pub fn get_total_swap(&self) -> Result<u64, Error>

Get the total amount of swap in the computer, in bytes.

Blocking

This is NOT blocking.

pub fn get_available_swap(&self) -> Result<u64, Error>

Get the available amount of swap in the computer, in bytes.

Blocking

This is NOT blocking.

pub fn count_logical_cpus(&self) -> Result<usize, Error>

Get the number of logical cpu cores.

Blocking

This is NOT blocking.